Bugfix: Allow loading first revision
Changing this selection (re)loads the page with the selected revision.
This is triggered only when the selection actually changes (on Firefox,
at least); if you just select the current revision, there is no reload.
This is good, since a reload would be unnecessary.

However, currently the first entry is always selected by default,
regardless of which revision is actually loaded / shown on page. This
makes it impossible to return to the first entry once you've left: in
order to do that, you'd need to switch off of it and then back on, but
switching off loads that revision and reloads the page with the first
entry selected again.

This commit changes the default selection to be the currently shown
revision, so it's possible to change back to the first one.